fix a node bug

+++ evaluation/node/first_run_gear_images.py
@@ -123,7 +123,8 @@ class Runner:
                 else:
                     time.sleep(5)
 
-                shutil.rmtree(localVolume)
+                if localVolume != "":
+                    shutil.rmtree(localVolume)
 
 class Generator:
